Hide user/team statistics if board is hidden
