+++ routes/board_routes.py
@@ -5,42 +5,14 @@ import random
 from flask import render_template, Blueprint, request, jsonify, make_response, redirect, url_for
 from flask_login import current_user
 import math
-from utils import autocomplete, database, db_entities, bingo, config, teampassword
+from utils import autocomplete, database, db_entities, bingo, board
 
 board_routes = Blueprint("board_routes", __name__)
 
-def is_board_globally_visible(current_user: object) -> bool:
-    """Determines if the board info is globally visible 
-
-    Args:
-        current_user (object): current user data
-
-    Returns:
-        bool: Whether the board should be shown
-    """
-    return config.allow_view_board(current_user.is_authenticated and current_user.is_admin)
-
-def is_team_authenticated(provided_pw: str, team_name: str) -> bool:
-    """Determines if the provided password is valid for a particular team ID
-
-    Args:
-        provided_pw (str): Password from URL
-        team_name (str): Team name
-
-    Returns:
-        bool: Whether this user is authorized to see this team's board
-    """
-    try:
-        team_entry = db_entities.Team(database.get_team_by_name(team_name)) 
-
-        return teampassword.calculate_team_password(team_entry.team_webhook) == provided_pw
-    except:
-        return False
-
 @board_routes.route('/compare', methods=['GET'])
 def compare():
-    if not is_board_globally_visible(current_user):
-        return hidden_board()
+    if not board.is_board_globally_visible(current_user):
+        return board.hidden_board()
 
     teams = []
     for team in database.get_teams():
@@ -78,8 +50,8 @@ class PanelData:
 
 @board_routes.route('/', methods=['GET'])
 def index():
-    if not is_board_globally_visible(current_user):
-        return hidden_board()
+    if not board.is_board_globally_visible(current_user):
+        return board.hidden_board()
 
     teams = []
     panelData = {}
@@ -134,36 +106,18 @@ def get_progress():
     return jsonify(progress.status_text)
 
 
-def select_random_file(directory):
-    # Get a list of all files in the directory
-    files = os.listdir(directory)
-
-    # Filter out directories, only keep files
-    files = [f for f in files if os.path.isfile(os.path.join(directory, f))]
-
-    # Select a random file
-    random_file = random.choice(files)
-
-    return random_file
-def hidden_board():
-
-    random_file = select_random_file("static/hidden_board_memes")
-    random_file = "hidden_board_memes/" + random_file
-    return render_template('board_templates/hidden_board.html', PageTitle="Board is hidden :)", FileName=random_file)
-
-
 @board_routes.route('/<team_name>', methods=['GET'])
 def board(team_name):
     team_password = request.args.get('pw') or None
     authenticated_team = ""
     if team_password is not None:
-        if not is_team_authenticated(team_password, team_name):
-            return hidden_board()
+        if not board.is_team_authenticated(team_password, team_name):
+            return board.hidden_board()
 
         authenticated_team = team_name
     else:
-        if not is_board_globally_visible(current_user):
-            return hidden_board()
+        if not board.is_board_globally_visible(current_user):
+            return board.hidden_board()
 
     panelData = {}
     tile_id_to_name = {}
